structure: move main file from pkg to cmd folder
