+++ .gitignore
@@ -2,4 +2,5 @@ config.yaml
 dist/
 *.pprof
 .vscode/
+.idea/
 /result
