+++ Dockerfile
@@ -14,13 +14,14 @@ COPY go.sum ./
 RUN go mod download
 
 COPY ./pkg ./pkg
+COPY ./cmd ./cmd
 
 # This runs a bit slower but guarantees that the binary does not rely on
 # the underlying C environment (e.g. as "static" as possible).
 RUN CGO_ENABLED=0 \
     GOOS=linux \
     GOARCH=amd64 \
-    go build -o /waterfall ./pkg
+    go build -o /waterfall ./cmd/sfu
 
 ##
 ## Deploy
