+++ pkg/profiling/profile.go
@@ -1,4 +1,4 @@
-package main
+package profiling
 
 import (
 	"os"
